django/requirements.txt
absl-py==0.14.1
asgiref==3.4.1
astunparse==1.6.3
bokeh==2.3.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
colorama==0.4.4
Django==3.2.8
flatbuffers==1.12
gast==0.4.0
google-api-core==2.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-vision==2.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.0
gTTS==2.2.3
h5py==3.1.0
idna==3.2
Jinja2==3.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.5
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
